+++ sys/ufs/ffs/ffs_rawread.c
@@ -192,7 +192,7 @@ ffs_rawread_readahead(struct vnode *vp,
 		      struct buf *bp)
 {
 	int error;
-	u_int iolen;
+	uint64_t iolen;
 	off_t blockno;
 	int blockoff;
 	int bsize;
@@ -264,7 +264,7 @@ ffs_rawread_main(struct vnode *vp,
 {
 	int error, nerror;
 	struct buf *bp, *nbp, *tbp;
-	u_int iolen;
+	uint64_t iolen;
 	caddr_t udata;
 	long resid;
 	off_t offset;
